Small styling changes with Checkout mostly completed
Check out successfully checks for remaining stock and prevents orders from completing if there are not enough items in stock. This is just in place since there isn't actually going to be a delivery person marking items as out of stock; the appropriate quantity is also taken from each item as well. The email service does not work quite yet, but it is set up in code.