Fixed email, Updated testing, and attempted automated searching
So emails are finally sending whenever an item is out of stock. I can use this for that receipt email to be sent to customers next week. The testing changes attempt to handle changes to the main controller made previously. The automated searching should be working in theory given the console logs, but it doesn't