Email receipt for customers completed.
Customers receive a receipt detailing their transactions. I need to work more on the email itself, but it provides each item, the price, the subtotal, and total of the purchase. I might also need to change how the email address is retrieved after this is moved to an experience cloud site as well.